- fixed (lat,long) error - updated some code for MapFragment - added getAllPostsCount() for the PostsDataSource - randomized the populated dummy posts

+++ Rove/src/com/fourkins/rove/sqlite/posts/PostsManager.java
@@ -1,6 +1,7 @@
 package com.fourkins.rove.sqlite.posts;
 
 import java.util.List;
+import java.util.Random;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -32,19 +33,27 @@ public class PostsManager {
     private void populateTable() {
         ds.open();
 
-        ds.deleteAllPosts();
+        if (ds.getAllPostsCount() == 0) {
+            ds.deleteAllPosts();
 
-        for (int i = 1; i <= 20; i++) {
-            ContentValues values = new ContentValues();
-            values.put(PostsSQLiteHelper.COLUMN_USER_NAME, "user" + i);
-            values.put(PostsSQLiteHelper.COLUMN_LATITUDE, i * 4.5);
-            values.put(PostsSQLiteHelper.COLUMN_LONGITUDE, i * 9);
-            values.put(PostsSQLiteHelper.COLUMN_MESSAGE, "message" + i);
+            Random rand = new Random();
 
-            ds.insertPost(values);
+            for (int i = 1; i <= 5; i++) {
+
+                // use 35,35 as the limit for now, keep things in Africa
+                int lat = rand.nextInt(35) * (rand.nextInt(2) == 0 ? -1 : 1);
+                int lng = rand.nextInt(35) * (rand.nextInt(2) == 0 ? -1 : 1);
+
+                ContentValues values = new ContentValues();
+                values.put(PostsSQLiteHelper.COLUMN_USER_NAME, "user" + i);
+                values.put(PostsSQLiteHelper.COLUMN_LATITUDE, lat);
+                values.put(PostsSQLiteHelper.COLUMN_LONGITUDE, lng);
+                values.put(PostsSQLiteHelper.COLUMN_MESSAGE, "message" + i);
+
+                ds.insertPost(values);
+            }
         }
 
         ds.close();
     }
-
 }
